separating session out into its own thing
